[Add] proper check for update-alternatives
We first check if there is a link group called
cmake and THEN we list the link group,
grepping for the specified version. Before,
the check would fail if there were no
update-alternatives installed on the system.
Such a scenario was common when an install
was being done without the self-bootstrap